+++ apps/web/src/_pages/console/OrganizationsPage/index.ts
@@ -1,2 +1,3 @@
 export { OrganizationsPage } from './OrganizationsPage'
 export { OrganizationsPageSkeleton } from './OrganizationsPageSkeleton'
+export { parsePage, parseSearch, parseSortBy, parseSortOrder } from './parse-query'